메인화면 공지사항, faq 카드 목록 추가
